[refactor] Consolidating aws related and Jenkins widget classes into their own packages.
