+++ src/main/java/com/amazon/jenkins/ec2fleet/aws/fleet/EC2SpotFleet.java
@@ -1,7 +1,7 @@
-package com.amazon.jenkins.ec2fleet.fleet;
+package com.amazon.jenkins.ec2fleet.aws.fleet;
 
 import com.amazon.jenkins.ec2fleet.FleetStateStats;
-import com.amazon.jenkins.ec2fleet.Registry;
+import com.amazon.jenkins.ec2fleet.aws.Registry;
 import com.amazonaws.services.ec2.AmazonEC2;
 import com.amazonaws.services.ec2.model.ActiveInstance;
 import com.amazonaws.services.ec2.model.BatchState;
