+++ src/test/java/com/amazon/jenkins/ec2fleet/aws/fleet/EC2SpotFleetTest.java
@@ -1,8 +1,8 @@
-package com.amazon.jenkins.ec2fleet.fleet;
+package com.amazon.jenkins.ec2fleet.aws.fleet;
 
-import com.amazon.jenkins.ec2fleet.EC2Api;
+import com.amazon.jenkins.ec2fleet.aws.EC2Api;
 import com.amazon.jenkins.ec2fleet.FleetStateStats;
-import com.amazon.jenkins.ec2fleet.Registry;
+import com.amazon.jenkins.ec2fleet.aws.Registry;
 import com.amazonaws.services.ec2.AmazonEC2;
 import com.amazonaws.services.ec2.model.ActiveInstance;
 import com.amazonaws.services.ec2.model.BatchState;
